requirements.txt
boto==2.48.0
certifi==2018.4.16
chardet==3.0.4
dj-database-url==0.5.0
Django==2.0.3
django-cities==0.5.0.6
django-model-utils==3.1.1
django-storages==1.6.5
gunicorn==19.7.1
idna==2.6
numpy==1.15.0
Pillow==5.1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2018.3
requests==2.18.4
six==1.11.0
stripe==1.82.1
swapper==1.1.0
tqdm==4.23.1
urllib3==1.22
